requirements.txt
# Core
python-dotenv==1.0.0
openai==1.54.0

# Slack Integration
slack-sdk==3.33.2
slack-bolt==1.21.2

# Confluence Integration
atlassian-python-api==3.41.0

# GitHub Integration
PyGithub==2.4.0

# Vector Database and RAG
chromadb==0.4.24
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
tiktoken==0.8.0

# Document Processing
beautifulsoup4==4.12.3
markdownify==0.13.1
pypdf==5.1.0

# Utilities
requests==2.32.3
urllib3==2.2.3

# Development
pytest==8.3.4
black==24.10.0
flake8==7.1.1
